requirements.txt
requests==2.22.0
pycryptodome==3.9.9
secrets==1.0.2
